Revert gmn_sampling and C function call
