+++ .Rbuildignore
@@ -8,3 +8,4 @@ LICENSE
 ^.*\.Rproj$
 ^\.Rproj\.user$
 qr.R
+gmn_sampling_degree.c
